requirements.txt
matplotlib==3.10.8
numpy==2.4.1
opencv_python_headless==4.12.0.88
Pillow==12.1.0
seaborn==0.13.2
streamlit==1.51.0
torch==2.8.0
torchvision==0.23.0
